backend/requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
PyPDF2==3.0.1
python-multipart==0.0.6
anthropic==0.40.0
python-dotenv==1.0.0